eliminati tutti i pezzi di codice commentati e corretti:

+++ KNN_Project/inc/vector_extraction.h
@@ -15,6 +15,6 @@
 
 float calculateMean(float vect[]);
 float calculateVar(float vect[], float mean);
-float calculateCorr(float vect1[], float vect2[], int med1, int med2, int stdDev1, int stdDev2);
+float calculateCorr(float vect1[], float vect2[], float med1, float med2, float stdDev1, float stdDev2);
 
 #endif /* VECTOR_EXTRACTION_H_ */
